check if file exists before unlink in tests, move song and track to data, added interval and test

+++ test/data/io/JSONFile.spec.js
@@ -18,7 +18,9 @@ describe('cl.data.io.JSONFile', () => {
         if (!discardFile) {
             discardFile = true;
         } else {
-            fs.unlinkSync(filepath);
+            if (fs.existsSync(filepath)) {
+                fs.unlinkSync(filepath);
+            }
         }
     });
